requirements.txt
requests==2.25.0
crontab==0.22.9
python-telegram-bot==13.0